+++ R/distinct.R
@@ -17,33 +17,33 @@
 #'   z = c("a", "a", "b"))
 #'
 #' example_dt %>%
-#'   dt_distinct()
+#'   distinct.()
 #'
 #' example_dt %>%
-#'   dt_distinct(z)
-dt_distinct <- function(.data, ...) {
-  UseMethod("dt_distinct")
+#'   distinct.(z)
+distinct. <- function(.data, ...) {
+  UseMethod("distinct.")
 }
 
 #' @export
-dt_distinct.tidytable <- function(.data, ...) {
+distinct..tidytable <- function(.data, ...) {
 
   dots <- enexprs(...)
 
   if (length(dots) == 0) {
     unique(.data)
   } else {
-    dots <- dots_selector(.data, ...)
-
-    .data %>%
-      dt_select(!!!dots) %>%
-      unique()
+    unique(dt_select(.data, ...))
   }
 }
 
 #' @export
-dt_distinct.data.frame <- function(.data, ...) {
+distinct..data.frame <- function(.data, ...) {
   .data <- as_tidytable(.data)
 
-  dt_distinct(.data, ...)
+  distinct.(.data, ...)
 }
+
+#' @export
+#' @rdname distinct.
+dt_distinct <- distinct.