+++ R/filter.R
@@ -5,36 +5,60 @@
 #'
 #' @param .data A data.frame or data.table
 #' @param ... Conditions to filter by
+#' @param by Columns to group by if filtering with a summary function
 #'
 #' @export
 #'
 #' @examples
 #' example_dt <- data.table::data.table(
 #'   a = c(1,2,3),
-#'   b = c(4,5,6))
+#'   b = c(4,5,6),
+#'   c = c("a","a","b"))
 #'
 #' example_dt %>%
-#'   dt_filter(a >= 2, b >= 4)
-dt_filter <- function(.data, ...) {
-  UseMethod("dt_filter")
+#'   filter.(a >= 2, b >= 4)
+#'
+#' example_dt %>%
+#'   filter.(b <= mean(b), by = c)
+filter. <- function(.data, ..., by = NULL) {
+  UseMethod("filter.")
 }
 
 #' @export
-dt_filter.tidytable <- function(.data, ...) {
+filter..tidytable <- function(.data, ..., by = NULL) {
 
   dots <- enexprs(...)
 
-  for (dot in dots) {
-    .data <- eval_tidy(expr(
-      .data[!!dot]
-    ))
+  by <- enexpr(by)
+
+  if (is.null(by)) {
+    for (dot in dots) {
+      .data <- eval_expr(
+        .data[!!dot]
+      )
+    }
+  } else {
+    by <- vec_selector_by(.data, !!by)
+
+    col_order <- names(.data)
+
+    .data <- eval_expr(
+      .data[, .SD[reduce(list(!!!dots), '&')], by = !!by]
+    )
+
+    setcolorder(.data, col_order)
   }
   .data
 }
 
 #' @export
-dt_filter.data.frame <- function(.data, ...) {
+filter..data.frame <- function(.data, ..., by = NULL) {
   .data <- as_tidytable(.data)
+  by <- enexpr(by)
 
-  dt_filter(.data, ...)
+  filter.(.data, ..., by = !!by)
 }
+
+#' @export
+#' @rdname filter.
+dt_filter <- filter.