+++ R/group_split.R
@@ -3,6 +3,8 @@
 #' @description
 #' Split data frame by groups. Returns a list
 #'
+#' Supports enhanced selection
+#'
 #' @param .data A data.frame or data.table
 #' @param ... Groups to split by
 #' @export
@@ -15,29 +17,34 @@
 #'   d = c("a","a","a","b","b"))
 #'
 #' test_df %>%
-#'   dt_group_split(c, d)
-dt_group_split <- function(.data, ...) {
-  UseMethod("dt_group_split")
+#'   group_split.(c, d)
+group_split. <- function(.data, ...) {
+  UseMethod("group_split.")
 }
 
 #' @export
-dt_group_split.tidytable <- function(.data, ...) {
+group_split..tidytable <- function(.data, ...) {
 
   dots <- enexprs(...)
 
   if (length(dots) == 0) {
     list(.data)
   } else {
-    dots <- dots_selector(.data, ...) %>%
-      as.character()
+    dots <- as.character(dots_selector(.data, ...))
 
     unname(split(.data, by = dots))
   }
 }
 
 #' @export
-dt_group_split.data.frame <- function(.data, ...) {
+group_split..data.frame <- function(.data, ...) {
   .data <- as_tidytable(.data)
 
-  dt_group_split(.data, ...)
+  group_split.(.data, ...)
 }
+
+#' @export
+#' @rdname group_split.
+dt_group_split <- group_split.
+
+