+++ R/pull.R
@@ -4,7 +4,7 @@
 #' Pull a single variable from a data.table as a vector.
 #'
 #' @param .data A data.frame or data.table
-#' @param var The column to pull from the data.table
+#' @param var The column to pull from the data.table. If NULL, pulls the last column.
 #'
 #' @export
 #'
@@ -14,26 +14,30 @@
 #'   y = c(4,5,6))
 #'
 #' example_dt %>%
-#'   dt_pull(y)
-dt_pull <- function(.data, var = NULL) {
-  UseMethod("dt_pull")
+#'   pull.(y)
+pull. <- function(.data, var = NULL) {
+  UseMethod("pull.")
 }
 
 #' @export
-dt_pull.tidytable <- function(.data, var = NULL) {
+pull..tidytable <- function(.data, var = NULL) {
 
   var <- enexpr(var)
-  if (is.null(var)) stop("var must be supplied")
+  if (is.null(var)) var <- sym(names(.data)[ncol(.data)])
 
-  eval_tidy(expr(
-    .data[, !!var]
-  ))
+  # Base R translation is faster than data.table
+  eval_expr(
+    '$'(.data, !!var)
+  )
 }
 
 #' @export
-dt_pull.data.frame <- function(.data, var = NULL) {
+pull..data.frame <- function(.data, var = NULL) {
   .data <- as_tidytable(.data)
   var <- enexpr(var)
 
-  dt_pull(.data, !!var)
+  pull.(.data, !!var)
 }
+#' @export
+#' @rdname pull.
+dt_pull <- pull.