+++ R/select_helpers.R
@@ -3,12 +3,14 @@
 #' @description
 #' These functions allow you to select variables based on their names.
 #'
-#' * `dt_starts_with()`: Starts with a prefix
-#' * `dt_starts_with()`: Ends with a suffix
-#' * `dt_contains()`: Contains a literal string
-#' * `dt_everything()`: Matches all variables
+#' * `any_of.()`: Select using a character vector
+#' * `contains.()`: Contains a literal string or regex match
+#' * `everything.()`: Matches all variables
+#' * `starts_with.()`: Starts with a prefix
+#' * `ends_with.()`: Ends with a suffix
 #'
 #' @param match The match for the helper function to use
+#' @param x Character vector of columns to select
 #'
 #' @md
 #' @export
@@ -21,44 +23,72 @@
 #'   double_y = 4)
 #'
 #' example_dt %>%
-#'   dt_select(dt_starts_with("x"))
+#'   select.(starts_with.("x"))
 #'
 #' example_dt %>%
-#'   dt_select(dt_ends_with("y"))
-dt_starts_with <- function(match) {
+#'   select.(ends_with.("y"))
+starts_with. <- function(match) {
   .names <- names(caller_env())
-  .names <- .names[!.names %in% c("is.integer", "is.double", "is.numeric",
-                                  "is.character", "is.factor", "is.logical")]
+  .names <- .names[!.names %in% predicate_names]
 
   seq_along(.names)[str_starts(.names, match)]
 }
 
 #' @export
-#' @rdname dt_starts_with
-dt_contains <- function(match) {
+#' @rdname starts_with.
+dt_starts_with <- starts_with.
+
+#' @export
+#' @rdname starts_with.
+contains. <- function(match) {
   .names <- names(caller_env())
-  .names <- .names[!.names %in% c("is.integer", "is.double", "is.numeric",
-                                  "is.character", "is.factor", "is.logical")]
+  .names <- .names[!.names %in% predicate_names]
 
   seq_along(.names)[str_detect(.names, match)]
 }
 
 #' @export
-#' @rdname dt_starts_with
-dt_ends_with <- function(match) {
+#' @rdname starts_with.
+dt_contains <- contains.
+
+#' @export
+#' @rdname starts_with.
+ends_with. <- function(match) {
   .names <- names(caller_env())
-  .names <- .names[!.names %in% c("is.integer", "is.double", "is.numeric",
-                                  "is.character", "is.factor", "is.logical")]
+  .names <- .names[!.names %in% predicate_names]
 
   seq_along(.names)[str_ends(.names, match)]
 }
 
 #' @export
-#' @rdname dt_starts_with
-dt_everything <- function() {
+#' @rdname starts_with.
+dt_ends_with <- ends_with.
+
+#' @export
+#' @rdname starts_with.
+everything. <- function() {
   .names <- names(caller_env())
-  .names <- .names[!.names %in% c("is.integer", "is.double", "is.numeric",
-                                  "is.character", "is.factor", "is.logical")]
+  .names <- .names[!.names %in% predicate_names]
 
   seq_along(.names)
 }
+
+#' @export
+#' @rdname starts_with.
+dt_everything <- everything.
+
+#' @export
+#' @rdname starts_with.
+any_of. <- function(x) {
+  .names <- names(caller_env())
+  .names <- .names[!.names %in% predicate_names]
+
+  seq_along(.names)[.names %in% x]
+}
+
+#' @export
+#' @rdname starts_with.
+dt_any_of <- any_of.
+
+predicate_names <- c("is.integer", "is.double", "is.numeric",
+                     "is.character", "is.factor", "is.logical", "is.list")